added lootin support, changed the command folder
