+++ src/main/java/fr/black_eyes/lootchest/Lootchest.java
@@ -413,7 +413,14 @@ public class Lootchest {
 			state.setData(data);
 			state.update();
 		}
-		
+		// check if lootin is installed
+		if(Config.getInstance().lootin && Bukkit.getPluginManager().isPluginEnabled("Lootin")) {
+			if(block.getType().equals(Material.CHEST) || block.getType().equals(Material.TRAPPED_CHEST))
+				com.github.sachin.lootin.utils.ChestUtils.setLootinContainer(null,block.getState(),com.github.sachin.lootin.utils.ContainerType.CHEST);
+			else if(block.getType().equals(Material.BARREL))
+				com.github.sachin.lootin.utils.ChestUtils.setLootinContainer(null,block.getState(),com.github.sachin.lootin.utils.ContainerType.BARREL);
+		}
+
 		// spawn particles and hologram if needed
 		final Location loc2 = getParticleLocation();
 		if(getParticle() != null && Main.configs.PART_enable){
