+++ src/main/java/fr/black_eyes/lootchest/commands/commands/CreateCommand.java
@@ -1,18 +1,22 @@
-package fr.black_eyes.lootchest.commands;
+package fr.black_eyes.lootchest.commands.commands;
 
-import fr.black_eyes.lootchest.Constants;
-import fr.black_eyes.lootchest.Lootchest;
-import fr.black_eyes.lootchest.Main;
-import fr.black_eyes.lootchest.Mat;
-import fr.black_eyes.lootchest.Utils;
-import fr.black_eyes.lootchest.ui.UiHandler;
 import java.util.Arrays;
+
 import org.bukkit.Material;
 import org.bukkit.block.Block;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.InventoryHolder;
 import org.bukkit.util.BlockIterator;
+
+import fr.black_eyes.lootchest.Constants;
+import fr.black_eyes.lootchest.Lootchest;
+import fr.black_eyes.lootchest.Main;
+import fr.black_eyes.lootchest.Mat;
+import fr.black_eyes.lootchest.Utils;
+import fr.black_eyes.lootchest.commands.ArgType;
+import fr.black_eyes.lootchest.commands.SubCommand;
+import fr.black_eyes.lootchest.ui.UiHandler;
 
 public class CreateCommand extends SubCommand {
