+++ src/main/java/fr/black_eyes/lootchest/commands/commands/GetNameCommand.java
@@ -1,13 +1,15 @@
-package fr.black_eyes.lootchest.commands;
+package fr.black_eyes.lootchest.commands.commands;
 
-import fr.black_eyes.lootchest.Constants;
-import fr.black_eyes.lootchest.Lootchest;
-import fr.black_eyes.lootchest.Utils;
 import org.bukkit.Material;
 import org.bukkit.block.Block;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 import org.bukkit.util.BlockIterator;
+
+import fr.black_eyes.lootchest.Constants;
+import fr.black_eyes.lootchest.Lootchest;
+import fr.black_eyes.lootchest.Utils;
+import fr.black_eyes.lootchest.commands.SubCommand;
 
 public class GetNameCommand extends SubCommand {
