+++ src/main/java/fr/black_eyes/lootchest/commands/commands/LocateCommand.java
@@ -1,11 +1,13 @@
-package fr.black_eyes.lootchest.commands;
+package fr.black_eyes.lootchest.commands.commands;
+
+import org.bukkit.Location;
+import org.bukkit.command.CommandSender;
 
 import fr.black_eyes.lootchest.Constants;
 import fr.black_eyes.lootchest.Lootchest;
 import fr.black_eyes.lootchest.Main;
 import fr.black_eyes.lootchest.Utils;
-import org.bukkit.Location;
-import org.bukkit.command.CommandSender;
+import fr.black_eyes.lootchest.commands.SubCommand;
 
 
 public class LocateCommand extends SubCommand {
