+++ src/main/java/fr/black_eyes/lootchest/commands/commands/ReloadCommand.java
@@ -1,12 +1,14 @@
-package fr.black_eyes.lootchest.commands;
+package fr.black_eyes.lootchest.commands.commands;
+
+import org.bukkit.Bukkit;
+import org.bukkit.command.CommandSender;
 
 import fr.black_eyes.lootchest.Config;
 import fr.black_eyes.lootchest.Files;
 import fr.black_eyes.lootchest.Lootchest;
 import fr.black_eyes.lootchest.Main;
 import fr.black_eyes.lootchest.Utils;
-import org.bukkit.Bukkit;
-import org.bukkit.command.CommandSender;
+import fr.black_eyes.lootchest.commands.SubCommand;
 
 public class ReloadCommand extends SubCommand {
