+++ src/main/java/fr/black_eyes/lootchest/commands/commands/TpCommand.java
@@ -1,14 +1,17 @@
-package fr.black_eyes.lootchest.commands;
+package fr.black_eyes.lootchest.commands.commands;
+
+import java.util.Arrays;
+
+import org.bukkit.Location;
+import org.bukkit.command.CommandSender;
+import org.bukkit.entity.Player;
 
 import fr.black_eyes.lootchest.Constants;
 import fr.black_eyes.lootchest.Lootchest;
 import fr.black_eyes.lootchest.Main;
 import fr.black_eyes.lootchest.Utils;
-import org.bukkit.Location;
-import org.bukkit.command.CommandSender;
-import org.bukkit.entity.Player;
-
-import java.util.Arrays;
+import fr.black_eyes.lootchest.commands.ArgType;
+import fr.black_eyes.lootchest.commands.SubCommand;
 
 public class TpCommand extends SubCommand {
