+++ src/main/resources/plugin.yml
@@ -3,7 +3,7 @@ main: fr.black_eyes.lootchest.Main
 version: 2.4.5
 author: Black_Eyes
 api-version: 1.13
-softdepend: [MultiWorld, Multiverse-Core, FabledSkyBlock, MultiverseCore, GriefPrevention, Residence, WorldGuard, Factions, FactionsX, Towny]
+softdepend: [MultiWorld, Multiverse-Core, FabledSkyBlock, MultiverseCore, GriefPrevention, Residence, WorldGuard, Factions, FactionsX, Towny, Lootin]
 description: >
              Plugin for Regenerating Chests
 commands:
